Clean up local sim class structure (#93)
