+++ test/unit_tests/braket/default_simulator/test_density_matrix_simulator.py
@@ -1,16 +1,3 @@
-# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"). You
-# may not use this file except in compliance with the License. A copy of
-# the License is located at
-#
-#     http://aws.amazon.com/apache2.0/
-#
-# or in the "license" file accompanying this file. This file is
-# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
-
 import cmath
 import json
 import sys
@@ -22,78 +9,91 @@ from braket.device_schema.simulators import (
     GateModelSimulatorDeviceCapabilities,
     GateModelSimulatorDeviceParameters,
 )
-from braket.ir.jaqcd import Program
+from braket.ir.jaqcd import Program as JaqcdProgram
+from braket.ir.openqasm import Program as OpenQASMProgram
 from braket.task_result import AdditionalMetadata, ResultTypeValue, TaskMetadata
 
-from braket.default_simulator.density_matrix_simulator import DensityMatrixSimulator
+from braket.default_simulator import DensityMatrixSimulator
 
 CircuitData = namedtuple("CircuitData", "circuit_ir probability_zero")
 
 
-invalid_ir_result_types = [
-    {"type": "statevector"},
-    {"type": "amplitude", "states": ["11"]},
-]
+@pytest.fixture(params=["OpenQASM", "Jaqcd"])
+def ir_type(request):
+    return request.param
 
 
 @pytest.fixture
 def noisy_circuit_2_qubit():
-    return Program.parse_raw(
-        json.dumps(
-            {
-                "instructions": [
-                    {"type": "x", "target": 0},
-                    {"type": "x", "target": 1},
-                    {"type": "bit_flip", "target": 1, "probability": 0.1},
-                ]
-            }
+    return (
+        JaqcdProgram.parse_raw(
+            json.dumps(
+                {
+                    "instructions": [
+                        {"type": "x", "target": 0},
+                        {"type": "x", "target": 1},
+                        {"type": "bit_flip", "target": 1, "probability": 0.1},
+                    ]
+                }
+            )
+        )
+        if ir_type == "Jaqcd"
+        else OpenQASMProgram(
+            source="""
+                OPENQASM 3.0;
+                qubit[2] q;
+
+                x q;
+                #pragma braket noise bit_flip(.1) q[1]
+                """
         )
     )
 
 
 @pytest.fixture
-def grcs_8_qubit():
-    with open("test/resources/grcs_8.json") as circuit_file:
-        data = json.load(circuit_file)
-        return CircuitData(Program.parse_raw(json.dumps(data["ir"])), data["probability_zero"])
+def grcs_8_qubit(ir_type):
+    if ir_type == "Jaqcd":
+        with open("test/resources/grcs_8.json") as circuit_file:
+            data = json.load(circuit_file)
+            return CircuitData(
+                JaqcdProgram.parse_raw(json.dumps(data["ir"])), data["probability_zero"]
+            )
+    return CircuitData(OpenQASMProgram(source="test/resources/grcs_8.qasm"), 0.0007324)
 
 
 @pytest.fixture
-def bell_ir():
-    return Program.parse_raw(
-        json.dumps(
-            {
-                "instructions": [
-                    {"type": "h", "target": 0},
-                    {"type": "cnot", "target": 1, "control": 0},
-                ]
-            }
-        )
-    )
-
-
-@pytest.fixture
-def bell_ir_with_result():
-    def _bell_ir_with_result(targets=None):
-        return Program.parse_raw(
+def bell_ir(ir_type):
+    return (
+        JaqcdProgram.parse_raw(
             json.dumps(
                 {
                     "instructions": [
                         {"type": "h", "target": 0},
                         {"type": "cnot", "target": 1, "control": 0},
-                    ],
-                    "results": [{"type": "expectation", "observable": ["x"], "targets": targets}],
+                    ]
                 }
             )
         )
+        if ir_type == "Jaqcd"
+        else OpenQASMProgram(
+            source="""
+            OPENQASM 3.0;
+            qubit[2] q;
 
-    return _bell_ir_with_result
+            h q[0];
+            cnot q[0], q[1];
+            """
+        )
+    )
 
 
 def test_simulator_run_noisy_circuit(noisy_circuit_2_qubit):
     simulator = DensityMatrixSimulator()
     shots_count = 10000
-    result = simulator.run(noisy_circuit_2_qubit, qubit_count=2, shots=shots_count)
+    if isinstance(noisy_circuit_2_qubit, JaqcdProgram):
+        result = simulator.run(noisy_circuit_2_qubit, qubit_count=2, shots=shots_count)
+    else:
+        result = simulator.run(noisy_circuit_2_qubit, shots=shots_count)
 
     assert all([len(measurement) == 2] for measurement in result.measurements)
     assert len(result.measurements) == shots_count
@@ -107,20 +107,13 @@ def test_simulator_run_noisy_circuit(noisy_circuit_2_qubit):
     assert result.additionalMetadata == AdditionalMetadata(action=noisy_circuit_2_qubit)
 
 
-@pytest.mark.parametrize("result_type", invalid_ir_result_types)
-@pytest.mark.xfail(raises=TypeError)
-def test_simulator_run_invalid_ir_result_types(result_type):
-    simulator = DensityMatrixSimulator()
-    ir = Program.parse_raw(
-        json.dumps({"instructions": [{"type": "h", "target": 0}], "results": [result_type]})
-    )
-    simulator.run(ir, qubit_count=2, shots=100)
-
-
 def test_simulator_run_bell_pair(bell_ir):
     simulator = DensityMatrixSimulator()
     shots_count = 10000
-    result = simulator.run(bell_ir, qubit_count=2, shots=shots_count)
+    if isinstance(bell_ir, JaqcdProgram):
+        result = simulator.run(bell_ir, qubit_count=2, shots=shots_count)
+    else:
+        result = simulator.run(bell_ir, shots=shots_count)
 
     assert all([len(measurement) == 2] for measurement in result.measurements)
     assert len(result.measurements) == shots_count
@@ -137,20 +130,256 @@ def test_simulator_run_bell_pair(bell_ir):
 @pytest.mark.xfail(raises=ValueError)
 def test_simulator_run_no_results_no_shots(bell_ir):
     simulator = DensityMatrixSimulator()
-    simulator.run(bell_ir, qubit_count=2, shots=0)
+    if isinstance(bell_ir, JaqcdProgram):
+        simulator.run(bell_ir, qubit_count=2, shots=0)
+    else:
+        simulator.run(bell_ir, shots=0)
 
 
 def test_simulator_run_grcs_8(grcs_8_qubit):
     simulator = DensityMatrixSimulator()
-    result = simulator.run(grcs_8_qubit.circuit_ir, qubit_count=8, shots=0)
+    if isinstance(grcs_8_qubit.circuit_ir, JaqcdProgram):
+        result = simulator.run(grcs_8_qubit.circuit_ir, qubit_count=8, shots=0)
+    else:
+        result = simulator.run(grcs_8_qubit.circuit_ir, shots=0)
     density_matrix = result.resultTypes[0].value
     assert cmath.isclose(density_matrix[0][0].real, grcs_8_qubit.probability_zero, abs_tol=1e-7)
+
+
+def test_properties():
+    simulator = DensityMatrixSimulator()
+    observables = ["x", "y", "z", "h", "i", "hermitian"]
+    max_shots = sys.maxsize
+    qubit_count = 13
+    expected_properties = GateModelSimulatorDeviceCapabilities.parse_obj(
+        {
+            "service": {
+                "executionWindows": [
+                    {
+                        "executionDay": "Everyday",
+                        "windowStartHour": "00:00",
+                        "windowEndHour": "23:59:59",
+                    }
+                ],
+                "shotsRange": [0, max_shots],
+            },
+            "action": {
+                "braket.ir.openqasm.program": {
+                    "actionType": "braket.ir.openqasm.program",
+                    "version": ["1"],
+                    "supportedOperations": sorted(
+                        [
+                            # OpenQASM primitives
+                            "U",
+                            "GPhase",
+                            # builtin Braket gates
+                            "ccnot",
+                            "cnot",
+                            "cphaseshift",
+                            "cphaseshift00",
+                            "cphaseshift01",
+                            "cphaseshift10",
+                            "cswap",
+                            "cv",
+                            "cy",
+                            "cz",
+                            "ecr",
+                            "h",
+                            "i",
+                            "iswap",
+                            "pswap",
+                            "phaseshift",
+                            "rx",
+                            "ry",
+                            "rz",
+                            "s",
+                            "si",
+                            "swap",
+                            "t",
+                            "ti",
+                            "unitary",
+                            "v",
+                            "vi",
+                            "x",
+                            "xx",
+                            "xy",
+                            "y",
+                            "yy",
+                            "z",
+                            "zz",
+                            # noise operations
+                            "bit_flip",
+                            "phase_flip",
+                            "pauli_channel",
+                            "depolarizing",
+                            "two_qubit_depolarizing",
+                            "two_qubit_dephasing",
+                            "amplitude_damping",
+                            "generalized_amplitude_damping",
+                            "phase_damping",
+                            "kraus",
+                        ]
+                    ),
+                    "supportedResultTypes": [
+                        {
+                            "name": "Sample",
+                            "observables": observables,
+                            "minShots": 1,
+                            "maxShots": max_shots,
+                        },
+                        {
+                            "name": "Expectation",
+                            "observables": observables,
+                            "minShots": 0,
+                            "maxShots": max_shots,
+                        },
+                        {
+                            "name": "Variance",
+                            "observables": observables,
+                            "minShots": 0,
+                            "maxShots": max_shots,
+                        },
+                        {"name": "Probability", "minShots": 0, "maxShots": max_shots},
+                        {"name": "DensityMatrix", "minShots": 0, "maxShots": 0},
+                    ],
+                },
+                "braket.ir.jaqcd.program": {
+                    "actionType": "braket.ir.jaqcd.program",
+                    "version": ["1"],
+                    "supportedOperations": [
+                        "amplitude_damping",
+                        "bit_flip",
+                        "ccnot",
+                        "cnot",
+                        "cphaseshift",
+                        "cphaseshift00",
+                        "cphaseshift01",
+                        "cphaseshift10",
+                        "cswap",
+                        "cv",
+                        "cy",
+                        "cz",
+                        "depolarizing",
+                        "ecr",
+                        "generalized_amplitude_damping",
+                        "h",
+                        "i",
+                        "iswap",
+                        "kraus",
+                        "pauli_channel",
+                        "two_qubit_pauli_channel",
+                        "phase_flip",
+                        "phase_damping",
+                        "phaseshift",
+                        "pswap",
+                        "rx",
+                        "ry",
+                        "rz",
+                        "s",
+                        "si",
+                        "swap",
+                        "t",
+                        "ti",
+                        "two_qubit_dephasing",
+                        "two_qubit_depolarizing",
+                        "unitary",
+                        "v",
+                        "vi",
+                        "x",
+                        "xx",
+                        "xy",
+                        "y",
+                        "yy",
+                        "z",
+                        "zz",
+                    ],
+                    "supportedResultTypes": [
+                        {
+                            "name": "Sample",
+                            "observables": observables,
+                            "minShots": 1,
+                            "maxShots": max_shots,
+                        },
+                        {
+                            "name": "Expectation",
+                            "observables": observables,
+                            "minShots": 0,
+                            "maxShots": max_shots,
+                        },
+                        {
+                            "name": "Variance",
+                            "observables": observables,
+                            "minShots": 0,
+                            "maxShots": max_shots,
+                        },
+                        {"name": "Probability", "minShots": 0, "maxShots": max_shots},
+                        {"name": "DensityMatrix", "minShots": 0, "maxShots": 0},
+                    ],
+                },
+            },
+            "paradigm": {"qubitCount": qubit_count},
+            "deviceParameters": GateModelSimulatorDeviceParameters.schema(),
+        }
+    )
+    assert simulator.properties == expected_properties
+
+
+def test_openqasm_density_matrix_simulator():
+    noisy_bell_qasm = """
+    qubit[2] qs;
+
+    h qs[0];
+    cnot qs[0], qs[1];
+
+    #pragma braket noise bit_flip(.2) qs[1]
+
+    #pragma braket result probability
+    """
+    device = DensityMatrixSimulator()
+    program = OpenQASMProgram(source=noisy_bell_qasm)
+    result = device.run(program)
+    probabilities = result.resultTypes[0].value
+    assert np.allclose(probabilities, [0.4, 0.1, 0.1, 0.4])
+
+
+invalid_ir_result_types = [
+    {"type": "statevector"},
+    {"type": "amplitude", "states": ["11"]},
+]
+
+
+@pytest.fixture
+def bell_ir_with_result():
+    def _bell_ir_with_result(targets=None):
+        return JaqcdProgram.parse_raw(
+            json.dumps(
+                {
+                    "instructions": [
+                        {"type": "h", "target": 0},
+                        {"type": "cnot", "target": 1, "control": 0},
+                    ],
+                    "results": [{"type": "expectation", "observable": ["x"], "targets": targets}],
+                }
+            )
+        )
+
+    return _bell_ir_with_result
+
+
+@pytest.mark.parametrize("result_type", invalid_ir_result_types)
+@pytest.mark.xfail(raises=TypeError)
+def test_simulator_run_invalid_ir_result_types(result_type):
+    simulator = DensityMatrixSimulator()
+    ir = JaqcdProgram.parse_raw(
+        json.dumps({"instructions": [{"type": "h", "target": 0}], "results": [result_type]})
+    )
+    simulator.run(ir, qubit_count=2, shots=100)
 
 
 @pytest.mark.xfail(raises=ValueError)
 def test_simulator_run_densitymatrix_shots():
     simulator = DensityMatrixSimulator()
-    ir = Program.parse_raw(
+    ir = JaqcdProgram.parse_raw(
         json.dumps(
             {"instructions": [{"type": "h", "target": 0}], "results": [{"type": "densitymatrix"}]}
         )
@@ -160,7 +389,7 @@ def test_simulator_run_densitymatrix_shots():
 
 def test_simulator_run_result_types_shots():
     simulator = DensityMatrixSimulator()
-    ir = Program.parse_raw(
+    ir = JaqcdProgram.parse_raw(
         json.dumps(
             {
                 "instructions": [
@@ -181,7 +410,7 @@ def test_simulator_run_result_types_shots():
 
 def test_simulator_run_result_types_shots_basis_rotation_gates():
     simulator = DensityMatrixSimulator()
-    ir = Program.parse_raw(
+    ir = JaqcdProgram.parse_raw(
         json.dumps(
             {
                 "instructions": [
@@ -204,7 +433,7 @@ def test_simulator_run_result_types_shots_basis_rotation_gates():
 @pytest.mark.xfail(raises=ValueError)
 def test_simulator_run_result_types_shots_basis_rotation_gates_value_error():
     simulator = DensityMatrixSimulator()
-    ir = Program.parse_raw(
+    ir = JaqcdProgram.parse_raw(
         json.dumps(
             {
                 "instructions": [
@@ -238,7 +467,7 @@ def test_simulator_bell_pair_result_types(bell_ir_with_result, targets):
 @pytest.mark.xfail(raises=ValueError)
 def test_simulator_fails_samples_0_shots():
     simulator = DensityMatrixSimulator()
-    prog = Program.parse_raw(
+    prog = JaqcdProgram.parse_raw(
         json.dumps(
             {
                 "instructions": [{"type": "h", "target": 0}],
@@ -312,7 +541,7 @@ def test_simulator_fails_samples_0_shots():
 )
 def test_simulator_valid_observables(result_types, expected):
     simulator = DensityMatrixSimulator()
-    prog = Program.parse_raw(
+    prog = JaqcdProgram.parse_raw(
         json.dumps(
             {
                 "instructions": [
@@ -326,102 +555,3 @@ def test_simulator_valid_observables(result_types, expected):
     result = simulator.run(prog, qubit_count=2, shots=0)
     for i in range(len(result_types)):
         assert np.allclose(result.resultTypes[i].value, expected[i])
-
-
-def test_properties():
-    simulator = DensityMatrixSimulator()
-    observables = ["x", "y", "z", "h", "i", "hermitian"]
-    max_shots = sys.maxsize
-    qubit_count = 13
-    expected_properties = GateModelSimulatorDeviceCapabilities.parse_obj(
-        {
-            "service": {
-                "executionWindows": [
-                    {
-                        "executionDay": "Everyday",
-                        "windowStartHour": "00:00",
-                        "windowEndHour": "23:59:59",
-                    }
-                ],
-                "shotsRange": [0, max_shots],
-            },
-            "action": {
-                "braket.ir.jaqcd.program": {
-                    "actionType": "braket.ir.jaqcd.program",
-                    "version": ["1"],
-                    "supportedOperations": [
-                        "amplitude_damping",
-                        "bit_flip",
-                        "ccnot",
-                        "cnot",
-                        "cphaseshift",
-                        "cphaseshift00",
-                        "cphaseshift01",
-                        "cphaseshift10",
-                        "cswap",
-                        "cv",
-                        "cy",
-                        "cz",
-                        "depolarizing",
-                        "ecr",
-                        "generalized_amplitude_damping",
-                        "h",
-                        "i",
-                        "iswap",
-                        "kraus",
-                        "pauli_channel",
-                        "two_qubit_pauli_channel",
-                        "phase_flip",
-                        "phase_damping",
-                        "phaseshift",
-                        "pswap",
-                        "rx",
-                        "ry",
-                        "rz",
-                        "s",
-                        "si",
-                        "swap",
-                        "t",
-                        "ti",
-                        "two_qubit_dephasing",
-                        "two_qubit_depolarizing",
-                        "unitary",
-                        "v",
-                        "vi",
-                        "x",
-                        "xx",
-                        "xy",
-                        "y",
-                        "yy",
-                        "z",
-                        "zz",
-                    ],
-                    "supportedResultTypes": [
-                        {
-                            "name": "Sample",
-                            "observables": observables,
-                            "minShots": 1,
-                            "maxShots": max_shots,
-                        },
-                        {
-                            "name": "Expectation",
-                            "observables": observables,
-                            "minShots": 0,
-                            "maxShots": max_shots,
-                        },
-                        {
-                            "name": "Variance",
-                            "observables": observables,
-                            "minShots": 0,
-                            "maxShots": max_shots,
-                        },
-                        {"name": "Probability", "minShots": 0, "maxShots": max_shots},
-                        {"name": "DensityMatrix", "minShots": 0, "maxShots": 0},
-                    ],
-                }
-            },
-            "paradigm": {"qubitCount": qubit_count},
-            "deviceParameters": GateModelSimulatorDeviceParameters.schema(),
-        }
-    )
-    assert simulator.properties == expected_properties
