+++ test/unit_tests/braket/default_simulator/test_simulator.py
@@ -15,8 +15,8 @@ import numpy as np
 import pytest
 
 from braket.default_simulator import observables
-from braket.default_simulator.jaqcd_simulator import BaseLocalJaqcdSimulator
 from braket.default_simulator.result_types import DensityMatrix, Expectation, Probability, Variance
+from braket.default_simulator.simulator import BaseLocalSimulator
 
 
 @pytest.mark.parametrize(
@@ -36,7 +36,7 @@ from braket.default_simulator.result_types import DensityMatrix, Expectation, Pr
     ],
 )
 def test_validate_result_types_qubits_exist(result_type):
-    BaseLocalJaqcdSimulator._validate_result_types_qubits_exist([result_type], 2)
+    BaseLocalSimulator._validate_result_types_qubits_exist([result_type], 2)
 
 
 @pytest.mark.xfail(raises=ValueError)
@@ -44,7 +44,7 @@ def test_validate_result_types_qubits_exist(result_type):
     "result_type", [Expectation(observables.PauliX([1])), DensityMatrix([1]), Probability([1])]
 )
 def test_validate_result_types_qubits_exist_error(result_type):
-    BaseLocalJaqcdSimulator._validate_result_types_qubits_exist([result_type], 1)
+    BaseLocalSimulator._validate_result_types_qubits_exist([result_type], 1)
 
 
 def test_observable_hash_tensor_product():
@@ -52,15 +52,15 @@ def test_observable_hash_tensor_product():
     obs = observables.TensorProduct(
         [observables.PauliX([0]), observables.Hermitian(matrix, [1, 2]), observables.PauliY([1])]
     )
-    hash_dict = BaseLocalJaqcdSimulator._observable_hash(obs)
+    hash_dict = BaseLocalSimulator._observable_hash(obs)
     matrix_hash = hash_dict[1]
     assert hash_dict == {0: "PauliX", 1: matrix_hash, 2: matrix_hash, 3: "PauliY"}
 
 
 def test_base_local_simulator_abstract():
     abstract_methods = (
-        "Can't instantiate abstract class BaseLocalJaqcdSimulator with "
+        "Can't instantiate abstract class BaseLocalSimulator with "
         "abstract methods initialize_simulation, properties"
     )
     with pytest.raises(TypeError, match=abstract_methods):
-        BaseLocalJaqcdSimulator()
+        BaseLocalSimulator()
